fix: organization member 분리 (#147)
