+++ core/network/src/main/java/com/easyhz/noffice/core/network/api/organization/OrganizationService.kt
@@ -2,9 +2,7 @@ package com.easyhz.noffice.core.network.api.organization
 
 import com.easyhz.noffice.core.network.model.request.organization.CategoryRequest
 import com.easyhz.noffice.core.network.model.request.organization.OrganizationCreationRequest
-import com.easyhz.noffice.core.network.model.request.organization.RegisterMemberRequest
 import com.easyhz.noffice.core.network.model.response.announcement.AnnouncementResponse
-import com.easyhz.noffice.core.network.model.response.announcement.MemberResponse
 import com.easyhz.noffice.core.network.model.response.category.CategoryUpdateResponse
 import com.easyhz.noffice.core.network.model.response.organization.OrganizationCapsuleResponse
 import com.easyhz.noffice.core.network.model.response.organization.OrganizationInformationResponse
@@ -15,7 +13,6 @@ import com.easyhz.noffice.core.network.util.NofficeResult
 import com.easyhz.noffice.core.network.util.PagingResult
 import retrofit2.http.Body
 import retrofit2.http.GET
-import retrofit2.http.PATCH
 import retrofit2.http.POST
 import retrofit2.http.PUT
 import retrofit2.http.Path
@@ -70,16 +67,4 @@ interface OrganizationService {
     suspend fun fetchOrganizationSignUpInfo(
         @Path("organizationId") organizationId: Int
     ): NofficeResult<OrganizationSignUpInformationResponse>
-
-    /* 조직 가입 대기자 조회 */
-    @GET("/api/v1/organizations/{organizationId}/pending-members")
-    suspend fun fetchOrganizationPendingMembers(
-        @Path("organizationId") organizationId: Int
-    ): NofficeResult<List<MemberResponse>>
-
-    @PATCH("/api/v1/organizations/{organizationId}/register")
-    suspend fun acceptRegisterMember(
-        @Path("organizationId") organizationId: Int,
-        @Body body: RegisterMemberRequest
-    ): NofficeResult<Unit>
 }
