+++ core/network/src/main/java/com/easyhz/noffice/core/network/di/service/OrganizationServiceModule.kt
@@ -1,5 +1,6 @@
 package com.easyhz.noffice.core.network.di.service
 
+import com.easyhz.noffice.core.network.api.organization.OrganizationMemberService
 import com.easyhz.noffice.core.network.api.organization.OrganizationService
 import com.easyhz.noffice.core.network.di.NofficeRetrofit
 import dagger.Module
@@ -14,4 +15,8 @@ object OrganizationServiceModule {
     @Provides
     fun provideOrganizationService(@NofficeRetrofit retrofit: Retrofit): OrganizationService =
         retrofit.create(OrganizationService::class.java)
+
+    @Provides
+    fun provideOrganizationMemberService(@NofficeRetrofit retrofit: Retrofit): OrganizationMemberService =
+        retrofit.create(OrganizationMemberService::class.java)
 }
