+++ data/organization/src/main/java/com/easyhz/noffice/data/organization/di/repository/OrganizationRepositoryModule.kt
@@ -1,5 +1,7 @@
 package com.easyhz.noffice.data.organization.di.repository
 
+import com.easyhz.noffice.data.organization.repository.member.OrganizationMemberRepository
+import com.easyhz.noffice.data.organization.repository.member.OrganizationMemberRepositoryImpl
 import com.easyhz.noffice.data.organization.repository.organization.OrganizationRepository
 import com.easyhz.noffice.data.organization.repository.organization.OrganizationRepositoryImpl
 import dagger.Binds
@@ -14,4 +16,9 @@ interface OrganizationRepositoryModule {
     fun bindOrganizationRepository(
         organizationRepositoryImpl: OrganizationRepositoryImpl
     ): OrganizationRepository
+
+    @Binds
+    fun bindOrganizationMemberRepository(
+        organizationMemberRepositoryImpl: OrganizationMemberRepositoryImpl
+    ): OrganizationMemberRepository
 }
