+++ data/organization/src/main/java/com/easyhz/noffice/data/organization/repository/member/OrganizationMemberRepository.kt
@@ -1,0 +1,9 @@
+package com.easyhz.noffice.data.organization.repository.member
+
+import com.easyhz.noffice.core.model.common.Member
+import com.easyhz.noffice.core.model.organization.param.RegisterMemberParam
+
+interface OrganizationMemberRepository {
+    suspend fun fetchOrganizationPendingMembers(organizationId: Int): Result<List<Member>>
+    suspend fun acceptRegisterMember(param: RegisterMemberParam): Result<Unit>
+}
