+++ data/organization/src/main/java/com/easyhz/noffice/data/organization/repository/organization/OrganizationRepository.kt
@@ -2,7 +2,6 @@ package com.easyhz.noffice.data.organization.repository.organization
 
 import androidx.paging.PagingData
 import com.easyhz.noffice.core.common.util.DateFormat
-import com.easyhz.noffice.core.model.common.Member
 import com.easyhz.noffice.core.model.organization.Organization
 import com.easyhz.noffice.core.model.organization.OrganizationInformation
 import com.easyhz.noffice.core.model.organization.OrganizationJoin
@@ -10,7 +9,6 @@ import com.easyhz.noffice.core.model.organization.OrganizationSignUpInformation
 import com.easyhz.noffice.core.model.organization.announcement.OrganizationAnnouncement
 import com.easyhz.noffice.core.model.organization.category.Category
 import com.easyhz.noffice.core.model.organization.param.OrganizationCreationParam
-import com.easyhz.noffice.core.model.organization.param.RegisterMemberParam
 import kotlinx.coroutines.flow.Flow
 
 interface OrganizationRepository {
@@ -21,6 +19,4 @@ interface OrganizationRepository {
     suspend fun updateOrganizationCategory(organizationId: Int, category: List<Int>): Result<List<Category>>
     suspend fun joinOrganization(organizationId: Int): Result<OrganizationJoin>
     suspend fun fetchOrganizationSignUpInfo(organizationId: Int): Result<OrganizationSignUpInformation>
-    suspend fun fetchOrganizationPendingMembers(organizationId: Int): Result<List<Member>>
-    suspend fun acceptRegisterMember(param: RegisterMemberParam): Result<Unit>
 }
