+++ domain/organization/src/main/java/com/easyhz/noffice/domain/organization/usecase/member/AcceptRegisterMemberUseCase.kt
@@ -1,14 +1,14 @@
-package com.easyhz.noffice.domain.organization.usecase.organization
+package com.easyhz.noffice.domain.organization.usecase.member
 
 import com.easyhz.noffice.core.common.base.BaseUseCase
 import com.easyhz.noffice.core.model.organization.param.RegisterMemberParam
-import com.easyhz.noffice.data.organization.repository.organization.OrganizationRepository
+import com.easyhz.noffice.data.organization.repository.member.OrganizationMemberRepository
 import javax.inject.Inject
 
 class AcceptRegisterMemberUseCase @Inject constructor(
-    private val organizationRepository: OrganizationRepository
+    private val organizationMemberRepository: OrganizationMemberRepository
 ): BaseUseCase<RegisterMemberParam, Unit>() {
     override suspend fun invoke(param: RegisterMemberParam): Result<Unit> {
-        return organizationRepository.acceptRegisterMember(param)
+        return organizationMemberRepository.acceptRegisterMember(param)
     }
 }
