+++ domain/organization/src/main/java/com/easyhz/noffice/domain/organization/usecase/member/FetchOrganizationPendingMembersUseCase.kt
@@ -1,14 +1,14 @@
-package com.easyhz.noffice.domain.organization.usecase.organization
+package com.easyhz.noffice.domain.organization.usecase.member
 
 import com.easyhz.noffice.core.common.base.BaseUseCase
 import com.easyhz.noffice.core.model.common.Member
-import com.easyhz.noffice.data.organization.repository.organization.OrganizationRepository
+import com.easyhz.noffice.data.organization.repository.member.OrganizationMemberRepository
 import javax.inject.Inject
 
 class FetchOrganizationPendingMembersUseCase @Inject constructor(
-    private val organizationRepository: OrganizationRepository
+    private val organizationMemberRepository: OrganizationMemberRepository
 ): BaseUseCase<Int, List<Member>>() {
     override suspend fun invoke(param: Int): Result<List<Member>> {
-        return organizationRepository.fetchOrganizationPendingMembers(param)
+        return organizationMemberRepository.fetchOrganizationPendingMembers(param)
     }
 }
